refactoring project to folders
