+++ src/test/java/school/redrover/FreestyleProjectTest.java
@@ -27,7 +27,7 @@ public class FreestyleProjectTest extends BaseTest {
     @Test(dependsOnMethods = "testCreateFreestyleProject")
     public void testDisableProject() {
         String warningProjectIsDisabled = new HomePage(getDriver())
-                .clickOnJobInListOfItems(PROJECT_NAME)
+                .clickOnJobInListOfItems(PROJECT_NAME, new FreestyleProjectPage(getDriver()))
                 .waitUntilTextNameProjectToBePresentInH1(PROJECT_NAME)
                 .clickConfigure()
                 .clickEnableDisableToggle()
@@ -40,7 +40,7 @@ public class FreestyleProjectTest extends BaseTest {
     @Test(dependsOnMethods = "testDisableProject")
     public void testEnableProject() {
         String projectIsEnabled = new HomePage(getDriver())
-                .clickOnJobInListOfItems(PROJECT_NAME)
+                .clickOnJobInListOfItems(PROJECT_NAME, new FreestyleProjectPage(getDriver()))
                 .waitUntilTextNameProjectToBePresentInH1(PROJECT_NAME)
                 .clickEnableButton()
                 .clickConfigure()
@@ -65,7 +65,7 @@ public class FreestyleProjectTest extends BaseTest {
         final String newProjectDescription = "This is a NEW freestyleProject description";
 
         FreestyleProjectPage freestyleProjectPage = new HomePage(getDriver())
-                .clickOnJobInListOfItems(PROJECT_NAME)
+                .clickOnJobInListOfItems(PROJECT_NAME, new FreestyleProjectPage(getDriver()))
                 .clickEditDescriptionButton()
                 .sendDescription(newProjectDescription)
                 .clickSave();
@@ -76,7 +76,7 @@ public class FreestyleProjectTest extends BaseTest {
     @Test(dependsOnMethods = "testEditDescription")
     public void testRenameFreestyleProject() {
         FreestyleProjectPage freestyleProjectPage = new HomePage(getDriver())
-                .clickOnJobInListOfItems(PROJECT_NAME)
+                .clickOnJobInListOfItems(PROJECT_NAME, new FreestyleProjectPage(getDriver()))
                 .clickLeftSideMenuRename()
                 .sendName(UPDATED_PROJECT_NAME)
                 .clickRename();
@@ -87,7 +87,7 @@ public class FreestyleProjectTest extends BaseTest {
     @Test(dependsOnMethods = "testRenameFreestyleProject")
     public void testDeleteFreestyleProject() {
         List<String> projectNameList = new HomePage(getDriver())
-                .clickOnJobInListOfItems(UPDATED_PROJECT_NAME)
+                .clickOnJobInListOfItems(UPDATED_PROJECT_NAME, new FreestyleProjectPage(getDriver()))
                 .clickLeftSideMenuDelete()
                 .clickPopUpYesDeleteProject()
                 .getProjectNameList();
